refactoring Schema::GetFieldIndex to return int

+++ cpp/src/arrow/type.cc
@@ -319,11 +319,11 @@ bool Schema::Equals(const Schema& other, bool check_metadata) const {
 }
 
 std::shared_ptr<Field> Schema::GetFieldByName(const std::string& name) const {
-  int64_t i = GetFieldIndex(name);
+  int i = GetFieldIndex(name);
   return i == -1 ? nullptr : fields_[i];
 }
 
-int64_t Schema::GetFieldIndex(const std::string& name) const {
+int Schema::GetFieldIndex(const std::string& name) const {
   if (fields_.size() > 0 && name_to_index_.size() == 0) {
     for (size_t i = 0; i < fields_.size(); ++i) {
       name_to_index_[fields_[i]->name()] = static_cast<int>(i);
